Create CasesAndStories.py
This tool helps you generate Test Cases and User Stories based on the info about the project that you provide. User is asked to add a name and description for the in-app screen that they're testing as well as all the elements (and description) on that screen. Using GPT3 model davinci-003, tool will generate list of test cases and ask you whether you want to generate user stories too.